test: update family test data
