+++ gcf_data_mapper/enums/family.py
@@ -1,0 +1,33 @@
+from enum import Enum
+
+
+class FamilyColumnsNames(Enum):
+    """The fields the GCF data mapper needs to parse family data/ metadata."""
+
+    APPROVED_REF = "ApprovedRef"
+    COUNTRIES = "Countries"
+    ENTITIES = "Entities"
+    FUNDING = "Funding"
+    PROJECT_URL = "ProjectURL"
+    PROJECTS_ID = "ProjectsID"
+    RESULT_AREAS = "ResultAreas"
+    SECTOR = "Sector"
+    THEME = "Theme"
+
+
+class FamilyNestedColumnNames(Enum):
+    """The fields the GCF data mapper needs to parse nested family data/ metadata."""
+
+    AREA = "Area"
+    BUDGET = "BudgetUSDeq"
+    NAME = "Name"
+    REGION = "Region"
+    SOURCE = "Source"
+    TYPE = "Type"
+
+
+class GCFProjectBudgetSource(Enum):
+    """The source of financing for the project's budget funding"""
+
+    CO_FINANCING = "Co-Financing"
+    GCF = "GCF"
